svelte-docker/django-docker/backend-django/pyproject.toml
[tool.poetry]
name = "django-board-backend-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2"
djangorestframework = "^3.15.1"
django-extensions = "^3.2.3"
django-cors-headers = "^4.3.1"
drf-spectacular = "^0.27.2"
drf-nested-routers = "^0.94.1"
django-filter = "^24.2"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
gunicorn = "^22.0.0"
djangorestframework-simplejwt = "^5.3.1"
django-stubs-ext = "^5.0.0"
dj-rest-auth = "^6.0.0"
requests = "^2.31.0"
cryptography = "^42.0.7"
django-allauth = "0.61.1"
pymysql = "^1.1.1"
django-mysql = "^4.14.0"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
pytest-playwright = "^0.5.1"
pytest-sugar = "^1.0.0"
pytest-mock = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
